requirements.txt
fastapi
uvicorn
praw
pandas
openpyxl
python-dotenv
crewai